Auth backend routes

+++ db/api/index.js
@@ -1,17 +1,29 @@
 import express from 'express'
 import 'dotenv/config'
 import { PrismaClient } from "@prisma/client"
-// import jwt from "jsonwebtoken"
+import jwt from "jsonwebtoken"
 const apiRouter = express.Router()
 const prisma = new PrismaClient()
 
-// apiRouter.use((req, res, next) => {
-//   const 
-// })
+// check for a valid token and set our req.user to the user in the token
+apiRouter.use((req, res, next) => {
+  const auth = req.headers.authorization
+  const token = auth?.startsWith("Bearer ") ? auth.slice(7) : null
+  try{
+    req.user = jwt.verify(token, process.env.JWT_SECRET)
+  }catch{
+    req.user = null
+  }
+  next()
+})
 
 // get a user by their id
 apiRouter.get('/finduser/:id', async (req, res, next) => {
+  console.log("req.user: ", req.user)
   const { id } = req.params
+  if(req.user.id !== id*1){
+    return res.send("Please login to do that")
+  }
   try{
     const user = await prisma.user.findUnique({
       where: {
@@ -25,37 +37,23 @@ apiRouter.get('/finduser/:id', async (req, res, next) => {
   }
 })
 
-// updating a high score (should authenticate we are correct user or admin)
-apiRouter.put('/finduser/:id', async (req, res, next) => {
-  const { id } = req.params
-  const { newHighscore } = req.body // might be a little tough to pass score data from the game
-  try{
-    const updatedHighscore = await prisma.user.update({
-      where: {
-        id: id*1,
-      },
-      data: {
-        highscore: newHighscore*1
-      }
-    })
-    res.send(updatedHighscore)
-  }catch (error){
-    console.log(error)
-    next(error)
-  }
-})
-
-// change username (should authenticate we are correct user or admin)
+// updating user info (authenticated for correct user and admin)
+// this query can be used to update highscore, username, or nickname, and it wont break if you leave the other keys blank
 apiRouter.put('/finduser/edit/:id', async (req, res, next) => {
   const { id } = req.params
-  const { newUsername } = req.body
+  const { newUsername, newHighscore, newNickname } = req.body
+  if(req.user.id !== id*1 && !req.user.isAdmin){
+    return res.send("You are not allowed to change that username")
+  }
   try{
     const updatedUser = await prisma.user.update({
       where: {
         id: id*1,
       },
       data: {
-        username: newUsername
+        username: newUsername,
+        highscore: newHighscore*1,
+        nickname: newNickname
       }
     })
     res.send(updatedUser)
@@ -67,6 +65,9 @@ apiRouter.put('/finduser/edit/:id', async (req, res, next) => {
 
 // find all users (for admins and for the leaderboard)
 apiRouter.get('/allusers', async (req, res, next) => {
+  if(!req.user){
+    return res.send("Please login to do that")
+  }
   try{
     const users = await prisma.user.findMany()
     res.send(users)
@@ -77,8 +78,33 @@ apiRouter.get('/allusers', async (req, res, next) => {
 })
 
 // create a new friend pair
-apiRouter.post('/friendpairs', async (req, res, next) => {
+// (authenticated for members of the friendpair)
+apiRouter.post('/friendpairs/create', async (req, res, next) => {
   const { friend1, friend2 } = req.body
+  if(req.user.id !== friend1*1 && req.user.id !== friend2*1){
+    return res.send("You can only make friendpairs that you are part of")
+  }
+
+  //make sure we dont already have a friendship between these users
+  const friendCheck = await prisma.friendPair.findFirst({
+    where: {
+      OR: [
+        {
+          friend1Id: friend1*1,
+          friend2Id: friend2*1
+        },
+        {
+          friend1Id: friend2*1,
+          friend2Id: friend1*1
+        }
+      ]
+    }    
+  })
+  
+  if(friendCheck){
+    return res.send("Users are already friends")
+  }
+
   try{
     const newFriendPair = await prisma.friendPair.create({
       data: {
@@ -109,10 +135,48 @@ apiRouter.post('/friendpairs/:friendpairid', async (req, res, next) => {
   }
 })
 
+// get all friendpairs that include designated user
+// (authenticated only for designated user and admins)
+apiRouter.get('/friendpairs/allfriends/:userid', async (req, res, next) => {
+  const { userid } = req.params
+  if(req.user.id !== userid*1 && !req.user.isAdmin){
+    return res.send("Must be signed in to get friends")
+  }
+  try{
+    const foundFriends = await prisma.friendPair.findMany({
+      where: {
+        OR: [
+          {
+            friend1Id: userid*1
+          },
+          {
+            friend2Id: userid*1
+          }
+        ]
+      }
+    })
+    res.send(foundFriends)
+  }catch (error){
+    console.log(error)
+    next(error)
+  }
+})
+
 // delete a friendpair </3
-// (should probably authenticate that we are the right user)
-apiRouter.delete('/friendpairs/:friendpairid', async (req, res, next) => {
+// (authenticated for members of friendpair)
+apiRouter.delete('/friendpairs/delete/:friendpairid', async (req, res, next) => {
   const { friendpairid } = req.params
+
+  const { friend1Id, friend2Id } = await prisma.friendPair.findUnique({
+    where: {
+      id: friendpairid*1
+    }
+  })
+
+  if(req.user.id !== friend1Id && req.user.id !== friend2Id ){
+    return res.send("You can only delete friendpairs that you are part of")
+  }
+
   try{ 
     const friendPairToDelete = await prisma.friendPair.delete({
       where: {
@@ -126,15 +190,18 @@ apiRouter.delete('/friendpairs/:friendpairid', async (req, res, next) => {
   }
 })
 
-// create a new message
+// create a new message (authenticated for message author)
 apiRouter.post('/friendpairs/chat/:chatid', async (req, res, next) => {
   const { chatid } = req.params
-  const { userId, content } = req.body
+  const { content } = req.body
+  if(!req.user.id){
+    return res.send("Must be signed in to send messages")
+  }
   try{
     const newMessage = await prisma.message.create({
       data: {
         chatId: chatid*1,
-        userId: userId*1,
+        userId: req.user.id,
         content: content
       }
     })
@@ -161,10 +228,18 @@ apiRouter.get('/friendpairs/chat/:chatid/allmessages', async (req, res, next) =>
   }
 })
 
-// update a message (should authenticate we are correct user or admin)
+// update a message (authenticated for message authors and admins)
 apiRouter.put('/friendpairs/chat/message/:messageid', async (req, res, next) => {
   const { messageid } = req.params
   const { newContent } = req.body
+  const { userId } = await prisma.message.findUnique({
+    where: {
+      id: messageid*1
+    }
+  })
+  if(req.user.id !== userId*1 && !req.user.isAdmin){
+    return res.send("User does not have permission to change that message")
+  }
   try{
     const updatedMessage = await prisma.message.update({
       where: {
@@ -181,9 +256,17 @@ apiRouter.put('/friendpairs/chat/message/:messageid', async (req, res, next) => 
   }
 })
 
-// delete a message
+// delete a message (authenticated for message authors and admins)
 apiRouter.delete('/friendpairs/chat/message/:messageid', async (req, res, next) => {
   const { messageid } = req.params
+  const { userId } = await prisma.message.findUnique({
+    where: {
+      id: messageid*1
+    }
+  })
+  if(req.user.id !== userId && !req.user.isAdmin){
+    return res.send("User does not have permission to delete that message")
+  }
   try{
     const deletedMessage = await prisma.message.delete({
       where: {
